refactor: 로그아웃 버튼 디자인 및 위치 수정
